Split out BlueprintContainer functionality further
