+++ osu.Game.Rulesets.Osu/Edit/OsuBlueprintContainer.cs
@@ -1,0 +1,35 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Game.Rulesets.Edit;
+using osu.Game.Rulesets.Objects.Drawables;
+using osu.Game.Rulesets.Osu.Edit.Blueprints.HitCircles;
+using osu.Game.Rulesets.Osu.Edit.Blueprints.Sliders;
+using osu.Game.Rulesets.Osu.Edit.Blueprints.Spinners;
+using osu.Game.Rulesets.Osu.Objects.Drawables;
+using osu.Game.Screens.Edit.Compose.Components;
+
+namespace osu.Game.Rulesets.Osu.Edit
+{
+    public class OsuBlueprintContainer : ComposeBlueprintContainer
+    {
+        public override SelectionHandler CreateSelectionHandler() => new OsuSelectionHandler();
+
+        public override SelectionBlueprint CreateBlueprintFor(DrawableHitObject hitObject)
+        {
+            switch (hitObject)
+            {
+                case DrawableHitCircle circle:
+                    return new HitCircleSelectionBlueprint(circle);
+
+                case DrawableSlider slider:
+                    return new SliderSelectionBlueprint(slider);
+
+                case DrawableSpinner spinner:
+                    return new SpinnerSelectionBlueprint(spinner);
+            }
+
+            return base.CreateBlueprintFor(hitObject);
+        }
+    }
+}
