+++ osu.Game/Screens/Edit/Compose/Components/ComposeBlueprintContainer.cs
@@ -1,0 +1,12 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+namespace osu.Game.Screens.Edit.Compose.Components
+{
+    /// <summary>
+    /// A blueprint container generally displayed as an overlay to a ruleset's playfield.
+    /// </summary>
+    public class ComposeBlueprintContainer : BlueprintContainer
+    {
+    }
+}
